2022-02-08 Sum digits until result is a single digit
Solve the problem posed by https://leetcode.com/problems/add-digits/.

Add the digits of a number until a sum is reached that is a single
digit. This problem is solved in two ways. Firstly the way I came up
with, via recursion, and secondly a more elegant solution done through
math knowledge of the Digital Root formula.